requirements.txt
django
psycopg2
channels
asgi_redis
celery
gunicorn
requests
rq
sorl-thumbnail
django-htmlmin
django-cachalot
django-redis
django-meta
django-gulp
python-cloudflare-v4
django-daterange-filter
django-webpack-loader
django_summernote
django-newsletter
Pillow
raven
